+++ packages/block-library/src/page-list/edit.js
@@ -27,8 +27,8 @@ import {
 } from '@wordpress/components';
 import { __, sprintf } from '@wordpress/i18n';
 import { useMemo, useState, useEffect, useCallback } from '@wordpress/element';
-import { useEntityRecords, store as coreStore } from '@wordpress/core-data';
-import { useSelect, useDispatch, select } from '@wordpress/data';
+import { useEntityRecords } from '@wordpress/core-data';
+import { useSelect, useDispatch } from '@wordpress/data';
 
 /**
  * Internal dependencies
@@ -258,7 +258,6 @@ export default function PageListEdit( {
 		hasDraggedChild,
 		isChildOfNavigation,
 	} = useSelect(
-		// eslint-disable-next-line no-shadow
 		( select ) => {
 			const {
 				getBlockParentsByBlockName,
@@ -407,57 +406,3 @@ export default function PageListEdit( {
 		</>
 	);
 }
-export const transforms = {
-	to: [
-		{
-			type: 'block',
-			blocks: [ 'core/list' ],
-			transform: () => {
-				const pages = select( coreStore ).getEntityRecords(
-					'postType',
-					'page',
-					{
-						per_page: -1,
-						status: 'publish',
-					}
-				);
-
-				if ( ! pages || pages.length === 0 ) {
-					return createBlock( 'core/list', {
-						ordered: false,
-					} );
-				}
-
-				const createListItems = ( parentId = 0 ) => {
-					return pages
-						.filter( ( page ) => page.parent === parentId )
-						.map( ( page ) => {
-							const childItems = createListItems( page.id );
-							const listItem = createBlock( 'core/list-item', {
-								content: `<a href="${ page.link }">${ page.title.rendered }</a>`,
-							} );
-							if ( childItems.length > 0 ) {
-								const subList = createBlock(
-									'core/list',
-									{},
-									childItems
-								);
-								listItem.innerBlocks = [ subList ];
-							}
-							return listItem;
-						} );
-				};
-
-				const innerBlocks = createListItems();
-
-				return createBlock(
-					'core/list',
-					{
-						ordered: false,
-					},
-					innerBlocks
-				);
-			},
-		},
-	],
-};
